Add refresh token

+++ src/utils/axios.js
@@ -31,10 +31,22 @@ axiosApiIntances.interceptors.response.use(
   function (error) {
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
+    console.log(error.response);
     if (error.response.status === 403) {
-      alert("Please Login !");
-      localStorage.clear();
-      window.location.href = "/login";
+      if (error.response.data.msg === "jwt expired") {
+        const refreshToken = localStorage.getItem("refreshToken");
+        axiosApiIntances
+          .post("auth/refresh", { refreshToken })
+          .then((res) => {
+            localStorage.setItem("token", res.data.data.token);
+            window.location.reload();
+          })
+          .catch((err) => console.log(err));
+      } else {
+        alert("Please Login !");
+        localStorage.clear();
+        window.location.href = "/login";
+      }
     }
     return Promise.reject(error);
   }
